extract AccountDict to its own module
